Fix it doesn't compile for x64.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2265 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/ui/condition.cpp
@@ -292,7 +292,7 @@ qm::SizeCondition::SizeCondition(const WCHAR* pwszName,
 								 UINT nValueId,
 								 UINT nDescriptionId,
 								 const WCHAR* pwszMacro,
-								 size_t nSize) :
+								 unsigned int nSize) :
 	Condition(pwszName),
 	nValueId_(nValueId),
 	nDescriptionId_(nDescriptionId),
